requirements.txt
Flask==2.1.2
firebase-rest-api